+++ sdk/keyvault/azure-keyvault-certificates/azure/keyvault/certificates/aio/client.py
@@ -511,7 +511,7 @@ class CertificateClient(AsyncKeyVaultClientBase):
         )
 
     @distributed_trace
-    def list_certificates(self, **kwargs: "**Any") -> AsyncIterable[CertificateProperties]:
+    def list_properties_of_certificates(self, **kwargs: "**Any") -> AsyncIterable[CertificateProperties]:
         """List certificates in the key vault.
 
         The GetCertificates operation returns the set of certificates resources
@@ -527,8 +527,8 @@ class CertificateClient(AsyncKeyVaultClientBase):
 
         Example:
             .. literalinclude:: ../tests/test_examples_certificates_async.py
-                :start-after: [START list_certificates]
-                :end-before: [END list_certificates]
+                :start-after: [START list_properties_of_certificates]
+                :end-before: [END list_properties_of_certificates]
                 :language: python
                 :caption: List all certificates
                 :dedent: 8
@@ -543,7 +543,9 @@ class CertificateClient(AsyncKeyVaultClientBase):
         )
 
     @distributed_trace
-    def list_certificate_versions(self, name: str, **kwargs: "**Any") -> AsyncIterable[CertificateProperties]:
+    def list_properties_of_certificate_versions(
+        self, name: str, **kwargs: "**Any"
+    ) -> AsyncIterable[CertificateProperties]:
         """List the versions of a certificate.
 
         The GetCertificateVersions operation returns the versions of a
@@ -558,8 +560,8 @@ class CertificateClient(AsyncKeyVaultClientBase):
 
         Example:
             .. literalinclude:: ../tests/test_examples_certificates_async.py
-                :start-after: [START list_certificate_versions]
-                :end-before: [END list_certificate_versions]
+                :start-after: [START list_properties_of_certificate_versions]
+                :end-before: [END list_properties_of_certificate_versions]
                 :language: python
                 :caption: List all versions of a certificate
                 :dedent: 8
@@ -947,8 +949,8 @@ class CertificateClient(AsyncKeyVaultClientBase):
         return CertificateIssuer._from_issuer_bundle(issuer_bundle=issuer_bundle)
 
     @distributed_trace
-    def list_issuers(self, **kwargs: "**Any") -> AsyncIterable[IssuerProperties]:
-        """List certificate issuers for the key vault.
+    def list_properties_of_issuers(self, **kwargs: "**Any") -> AsyncIterable[IssuerProperties]:
+        """Lists properties of the certificate issuers for the key vault.
 
         Returns the set of certificate issuer resources in the key
         vault. This operation requires the certificates/manageissuers/getissuers
@@ -960,8 +962,8 @@ class CertificateClient(AsyncKeyVaultClientBase):
 
         Example:
             .. literalinclude:: ../tests/test_examples_certificates_async.py
-                :start-after: [START list_issuers]
-                :end-before: [END list_issuers]
+                :start-after: [START list_properties_of_issuers]
+                :end-before: [END list_properties_of_issuers]
                 :language: python
                 :caption: List issuers of a vault
                 :dedent: 8