+++ sdk/keyvault/azure-keyvault-secrets/HISTORY.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 4.1.0b1 Unreleased
+
+
 ## 2019-10-31 4.0.0
 ### Breaking changes:
 - Moved optional parameters of two methods into kwargs (